+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.3.0
 
 * added automatic use of [errorhandler](https://www.npmjs.com/package/errorhandler) in `ApiHost` class, if in "local dev mode" (`IS_LOCAL_DEV` === `(true)`)
+* added `useErrorHandler()` to `ApiHost` class
 
 ## 2.2.0